// client/src/app/stores/orders/order.types.ts
export enum ORDERS_ACTION_TYPES {
  // Fetch all
    FETCH_ALL_ORDERS_LOADING = "orders/FETCH_ALL_ORDERS_LOADING",
    FETCH_ALL_ORDERS_SUCCESS = "orders/FETCH_ALL_ORDERS_SUCCESS",
    FETCH_ALL_ORDERS_FAILED = "orders/FETCH_ALL_ORDERS_FAILED",
    // Fetch by ID
    FETCH_ORDER_BY_ID_LOADING = "orders/FETCH_ORDER_BY_ID_LOADING",
    FETCH_ORDER_BY_ID_SUCCESS = "orders/FETCH_ORDER_BY_ID_SUCCESS",
    FETCH_ORDER_BY_ID_FAILED = "orders/FETCH_ORDER_BY_ID_FAILED",
    // Add Order
    ADD_ORDER_LOADING = "orders/ADD_ORDER_LOADING",
    ADD_ORDER_SUCCESS = "orders/ADD_ORDER_SUCCESS",
    ADD_ORDER_FAILED = "orders/ADD_ORDER_FAILED",
    // Update Order
    UPDATE_ORDER_LOADING = "orders/UPDATE_ORDER_LOADING",
    UPDATE_ORDER_SUCCESS = "orders/UPDATE_ORDER_SUCCESS",
    UPDATE_ORDER_FAILED = "orders/UPDATE_ORDER_FAILED",
    // Delete Order
    DELETE_ORDER_LOADING = "orders/DELETE_ORDER_LOADING",
    DELETE_ORDER_SUCCESS = "orders/DELETE_ORDER_SUCCESS",
    DELETE_ORDER_FAILED = "orders/DELETE_ORDER_FAILED",
  };
  